+++ src/main/java/kitchenpos/menu/application/MenuService.java
@@ -3,16 +3,16 @@ package kitchenpos.menu.application;
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.application.dto.MenuProductRequest;
-import kitchenpos.application.dto.MenuProductRequest.Create;
+import kitchenpos.menu.application.dto.MenuProductRequest;
+import kitchenpos.menu.application.dto.MenuProductRequest.Create;
 import kitchenpos.menu.application.dto.MenuRequest;
 import kitchenpos.menu.application.dto.MenuResponse;
 import kitchenpos.menu.domain.repository.MenuDao;
 import kitchenpos.menu.domain.repository.MenuGroupDao;
-import kitchenpos.dao.MenuProductDao;
+import kitchenpos.menu.domain.repository.MenuProductDao;
 import kitchenpos.product.domain.repository.ProductDao;
 import kitchenpos.menu.domain.Menu;
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.product.domain.Product;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
