+++ src/main/java/kitchenpos/menu/application/dto/MenuResponse.java
@@ -3,7 +3,7 @@ package kitchenpos.menu.application.dto;
 import java.math.BigDecimal;
 import java.util.List;
 import kitchenpos.menu.domain.Menu;
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.menu.domain.MenuProduct;
 
 public class MenuResponse {
