+++ src/main/java/kitchenpos/menu/domain/repository/MenuProductDao.java
@@ -1,8 +1,8 @@
-package kitchenpos.dao;
+package kitchenpos.menu.domain.repository;
 
 import java.util.List;
 import java.util.Optional;
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.menu.domain.MenuProduct;
 
 public interface MenuProductDao {
     MenuProduct save(MenuProduct entity);
