+++ src/main/java/kitchenpos/order/application/OrderService.java
@@ -2,16 +2,16 @@ package kitchenpos.order.application;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.application.dto.OrderLineItemRequest;
-import kitchenpos.application.dto.OrderLineItemRequest.Create;
+import kitchenpos.order.application.dto.OrderLineItemRequest;
+import kitchenpos.order.application.dto.OrderLineItemRequest.Create;
 import kitchenpos.order.application.dto.OrderRequest;
 import kitchenpos.order.application.dto.OrderResponse;
 import kitchenpos.menu.domain.repository.MenuDao;
 import kitchenpos.order.domain.repository.OrderDao;
-import kitchenpos.dao.OrderLineItemDao;
+import kitchenpos.order.domain.repository.OrderLineItemDao;
 import kitchenpos.dao.OrderTableDao;
 import kitchenpos.order.domain.Order;
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.domain.OrderTable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
