+++ src/main/java/kitchenpos/order/application/dto/OrderResponse.java
@@ -3,7 +3,7 @@ package kitchenpos.order.application.dto;
 import java.time.LocalDateTime;
 import java.util.List;
 import kitchenpos.order.domain.Order;
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderStatus;
 
 public class OrderResponse {
