+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -7,20 +7,20 @@ import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Collections;
 import java.util.List;
-import kitchenpos.application.dto.OrderLineItemRequest;
+import kitchenpos.order.application.dto.OrderLineItemRequest;
 import kitchenpos.order.application.dto.OrderRequest;
 import kitchenpos.order.application.dto.OrderRequest.Status;
 import kitchenpos.order.application.dto.OrderResponse;
 import kitchenpos.menu.domain.repository.MenuDao;
 import kitchenpos.menu.domain.repository.MenuGroupDao;
 import kitchenpos.order.domain.repository.OrderDao;
-import kitchenpos.dao.OrderLineItemDao;
+import kitchenpos.order.domain.repository.OrderLineItemDao;
 import kitchenpos.dao.OrderTableDao;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.order.application.OrderService;
 import kitchenpos.order.domain.Order;
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.domain.OrderTable;
 import kitchenpos.support.ServiceTest;
